Extract shared Traefik/Vite setup
