+++ packages/host/scripts/serve-dist.js
@@ -1,66 +1,16 @@
 /**
- * Wrapper around `vite preview` that supports dynamic port allocation in environment mode.
- * When BOXEL_ENVIRONMENT is set, picks a free port, starts preview, then registers
- * with Traefik so that `host.<branch>.localhost` routes to this instance.
- * When BOXEL_ENVIRONMENT is not set, previews on port 4200.
+ * Wrapper around `vite preview` for `pnpm serve:dist`. Delegates to the
+ * shared launcher, which handles BOXEL_ENVIRONMENT / Traefik registration.
+ * Mirrors scripts/vite-serve.js, which does the same for the dev server.
  *
- * CORS headers and SPA fallback are configured in vite.config.mjs under `preview`.
+ * CORS headers and SPA fallback are configured in vite.config.mjs under
+ * `preview`.
  */
 
-const { spawn } = require('child_process');
-const path = require('path');
+const { startWithTraefik } = require('./vite-with-traefik');
 
-const BOXEL_ENVIRONMENT = process.env.BOXEL_ENVIRONMENT;
-
-function runServe(port) {
-  const child = spawn(
-    'npx',
-    ['vite', 'preview', '--port', String(port), '--strictPort'],
-    { stdio: 'inherit', cwd: path.join(__dirname, '..'), shell: true },
-  );
-  child.on('exit', (code) => process.exit(code || 0));
-  return child;
-}
-
-if (!BOXEL_ENVIRONMENT) {
-  // Standard mode: hardcoded port 4200
-  runServe(4200);
-} else {
-  const { ensureTraefik } = require('./ensure-traefik');
-  const { getEnvSlug, registerWithTraefik } = require('./traefik-helpers');
-
-  ensureTraefik();
-
-  const net = require('net');
-
-  const slug = getEnvSlug();
-  const hostname = `host.${slug}.localhost`;
-
-  // Find a free port
-  const srv = net.createServer();
-  srv.listen(0, () => {
-    const port = srv.address().port;
-    srv.close(() => {
-      console.log(
-        `[environment-mode] Starting host app on dynamic port ${port}`,
-      );
-      console.log(
-        `[environment-mode] Will be accessible at http://${hostname}`,
-      );
-
-      runServe(port);
-
-      try {
-        registerWithTraefik(slug, hostname, port);
-        console.log(
-          `[environment-mode] Registered host at ${hostname} -> localhost:${port}`,
-        );
-      } catch (e) {
-        console.error(
-          '[environment-mode] Failed to register with Traefik:',
-          e.message,
-        );
-      }
-    });
-  });
-}
+startWithTraefik({
+  subcommand: 'preview',
+  defaultPort: 4200,
+  label: 'vite preview',
+});
